[Feature] Controller Setting
