+++ build.gradle
@@ -1,6 +1,7 @@
 buildscript {
     ext {
         compose_version = '1.3.0-beta01'
+        kotlin_version = '1.7.10'
     }
     repositories {
         google()
