+++ eupi-ppt-controller/src/main/AndroidManifest.xml
@@ -12,6 +12,11 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Waveppt">
         <activity
+            android:name=".ControllerActivity"
+            android:exported="false"
+            android:label="@string/title_activity_controller"
+            android:theme="@style/Theme.Waveppt" />
+        <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
