+++ eupi-ppt-controller/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">eupi-ppt-controller</string>
+    <string name="title_activity_controller">ControllerActivity</string>
 </resources>
