+++ settings.gradle
@@ -15,3 +15,4 @@ dependencyResolutionManagement {
 rootProject.name = "wave-ppt"
 include ':eupi-ppt-viewer'
 include ':eupi-ppt-controller'
+include ':euphony'
