--- wave-ppt/eupi-ppt-controller/src/main/java/com/euphony/eupi_ppt_controller/EuPICodeEnum.kt
@@ -1,6 +1,0 @@
-package com.euphony.eupi_ppt_controller
-
-enum class EuPICodeEnum(val code:Double) {
-    PREV_PAGE(18500.0),
-    NEXT_PAGE(20500.0)
-}
