--- wave-ppt/eupi-ppt-viewer/src/main/java/com/euphony/eupi_ppt_viewer/EuPICodeEnum.kt
@@ -1,6 +1,0 @@
-package values
-
-enum class EuPICodeEnum(val code:Double) {
-    PREV_PAGE(18500.0),
-    NEXT_PAGE(20500.0)
-}
